Offer the style engine's own shape fill to look and feel delegates
The style engine fills its surfaces through a routine which keeps antialiasing
only where a shape actually curves: a rounded rectangle becomes three plain
bands and four antialiased corner boxes, and an integer rectangle needs no
coverage arithmetic at all. It writes exactly the pixels a plain fill writes,
because it only takes the shortcut when every cut between an antialiased part
and a plain one lands on a whole device pixel.

A look and feel delegate painting its own chrome has the same shape and the
same problem, and no cache to hide it behind - Swing draws a scroll bar's thumb
and a progress bar's fill from scratch on every repaint. So the routine moves
out of StyleRenderer into swingtree.api.laf.ShapeRendering, next to the
interface a delegate already implements to be styled by the engine. There is
now one implementation rather than one plus whatever each look and feel writes
for itself.

No behaviour change: the ten showcase presets render byte-identically at two
widths.
