+++ src/main/java/swingtree/api/laf/package-info.java
@@ -1,0 +1,9 @@
+/**
+ *  What a custom <i>Look and Feel</i> needs in order to render through the SwingTree style engine:
+ *  the {@link swingtree.api.laf.SwingTreeStyledComponentUI} interface a
+ *  {@link javax.swing.plaf.ComponentUI} implements to be styled by it, and the drawing primitives
+ *  the engine itself uses, which a delegate painting its own chrome wants for the same reasons.
+ */
+@NullMarked package swingtree.api.laf;
+
+import org.jspecify.annotations.NullMarked;
